@@ -1,19 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-
 extension StringExtension on String {
-  bool validateEmail() {
-    return RegExp(
-            r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$')
-        .hasMatch(this);
-  }
-  
   bool validateName() {
     return RegExp(r'^([a-z A-Z,.\-]).{2,24}$').hasMatch(this);
   }
 }
-
 
 void showMessage(BuildContext context, String title, String message,
     [void Function()? onPressed]) {
@@ -29,20 +21,20 @@ void showMessage(BuildContext context, String title, String message,
             content: Text(message, textAlign: TextAlign.center),
             actions: <Widget>[
               ElevatedButton(
-                child: Text(
-                  'Okay!',
-                  style: Theme.of(context).textTheme.button,
-                ),
                 onPressed: onPressed ??
                     () {
                       Navigator.of(context).pop();
                     },
                 style: ElevatedButton.styleFrom(
                   minimumSize: const Size(double.infinity, 56.0),
+                  backgroundColor: Colors.orange,
                   shape: RoundedRectangleBorder(
                     borderRadius: BorderRadius.circular(12.0),
                   ),
-                  primary: Colors.orange,
+                ),
+                child: Text(
+                  'Okay!',
+                  style: Theme.of(context).textTheme.labelLarge,
                 ),
               )
             ],
@@ -50,7 +42,7 @@ void showMessage(BuildContext context, String title, String message,
 }
 
 void launchURL(BuildContext context, String url) async {
-  if (!await launch(url)) {
-    showMessage(context, 'Error', 'Could not launch $url');
-  }
+  await launchUrl(
+    Uri.parse(url),
+  );
 }
