@@ -1,19 +1,29 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 import 'package:me/bloc/logging/app_logging.dart';
 import 'package:me/bloc/message/message_cubit.dart';
 import 'package:me/bloc/project/project_cubit.dart';
 import 'package:me/data/repositories/api_repository.dart';
 import 'package:me/data/utils/contants.dart';
-import 'package:url_strategy/url_strategy.dart';
 import 'ui/pages/home_page.dart';
 
 Future<void> main() async {
-  setPathUrlStrategy();
-  BlocOverrides.runZoned(
-    () => runApp(const MyApp()),
-    blocObserver: Logging(),
-  );
+  /// initialize flutter binding
+  WidgetsFlutterBinding.ensureInitialized();
+
+  /// register license assets/fonts/OFL.txt
+  LicenseRegistry.addLicense(() async* {
+    final license = await rootBundle.loadString('assets/fonts/OFL.txt');
+    yield LicenseEntryWithLineBreaks(['urbanist'], license);
+  });
+
+  /// use web plugin for removing # from url
+  setUrlStrategy(PathUrlStrategy());
+  Bloc.observer = Logging();
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
@@ -33,18 +43,28 @@ class MyApp extends StatelessWidget {
           debugShowCheckedModeBanner: false,
           theme: ThemeData(
             brightness: Brightness.dark,
-            primarySwatch: Colors.orange,
             fontFamily: kFontFamily,
             useMaterial3: true,
-            backgroundColor: Colors.black,
             scaffoldBackgroundColor: Colors.black,
+            inputDecorationTheme: const InputDecorationTheme(
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.all(Radius.circular(16.0)),
+              ),
+            ),
             appBarTheme: const AppBarTheme(
-                elevation: 0.0, backgroundColor: Colors.transparent),
+              elevation: 0.0,
+              backgroundColor: Colors.transparent,
+            ),
+            colorScheme: ColorScheme.fromSwatch(
+              primarySwatch: Colors.orange,
+              brightness: Brightness.dark,
+            ).copyWith(
+              background: Colors.black,
+            ),
           ),
           home: const MyHomePage(),
         ),
       );
 }
-
 
 //https://script.google.com/macros/s/AKfycbybg2YLApS964dNOB8UJE7K3bKRA2obfIBxYGwDY7zkAV2ONr2-/exec
